test: move the Pro-helpers-locator workaround out of the vendored PythonMockSdk
PythonMockSdk is a vendored copy of a JetBrains class, and the 70-line
workaround added to it was 70 lines to re-merge the next time it is re-vendored.
It is a test-application bootstrap concern, not a mock-SDK one; only the
one-line call stays, because create() really is the point every test path
funnels through.

While moving it, latch on a *successful* removal again rather than on the
attempt, so a repopulated EP is handled, and give the "did nothing" warning its
own flag so a platform without the Pro locator retries quietly instead of
logging the same four lines once per scenario. Both flags are @Volatile:
create() is reached from the cucumber glue and from PyLightProjectDescriptor.

Also drops the MultiMap that was built and never read, which was causing
createRoots -- a typeshed scan plus two VFS refreshes -- to run twice per SDK.
